Update README.md
This project is more than just code—it’s a complete journey into the fascinating world of forecasting. Whether you're an aspiring data scientist, a seasoned analyst, or simply a curious mind, this repository will empower you with essential tools and techniques to unlock the hidden stories in your data.